requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-community>=0.0.16
langchain_huggingface~=0.2.0
langchain-groq>=0.1.1
langchain-openai>=0.0.5
langchain-google-genai>=0.0.3
langchain-ollama>=0.3.3
chromadb>=0.4.22
sentence-transformers>=2.2.2
pypdf2>=3.0.1
python-dotenv>=1.0.0
pyyaml>=6.0.1
tiktoken>=0.5.2
groq>=0.4.1
openai>=1.3.0
google-generativeai>=0.3.0
requests>=2.31.0
